refactor(arcade): isolate blackjack textures
